# docs/swagger.yaml
basePath: /
definitions:
  errors.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  handlers.authenticationRequest:
    properties:
      params:
        additionalProperties: true
        type: object
    type: object
...
  media.DownloadMetaData:
    properties:
      added_by:
        type: string
      alt_text:
        type: string
      created_at:
        type: string
      filename:
        type: string
      format:
        type: string
      height:
        type: integer
      id:
        type: string
      length:
        type: integer
      media_type:
        $ref: '#/definitions/media.MediaType'
      post_id:
        type: string
      size:
        description: in bytes
        type: integer
      url:
        type: string
      url_thumbnail:
        type: string
      width:
        type: integer
    type: object
  media.MediaType:
    enum:
    - image
    - video
    - short_video
    - document
    type: string
    x-enum-varnames:
    - MediaTypeImage
    - MediaTypeVideo
    - MediaTypeShortVideo
    - MediaTypeDocument
...
  post.PostType:
    enum:
    - text
    - mix_media
    - image
    - multi_image
    - video
    - short_video
    - document
    - carousel
    type: string
    x-enum-varnames:
    - PostTypeText
    - PostTypeMixMedia
    - PostTypeImage
    - PostTypeMultiImage
    - PostTypeVideo
    - PostTypeShortVideo
    - PostTypeDocument
    - PostTypeCarousel
  post.PublishPost:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_idea:
        type: boolean
      platform:
        type: string
      project_id:
        type: string
      publish_status:
        type: string
      scheduled_at:
        type: string
      secrets:
        description: Additional fields
        type: string
      status:
        type: string
      text_content:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/post.PostType'
      updated_at:
        type: string
    type: object
...
  publisher.Platform:
    properties:
      id:
        $ref: '#/definitions/publisher.PlatformID'
      name:
        type: string
    type: object
  publisher.PlatformID:
    enum:
    - facebook
    - x
    - linkedin
    - instagram
    type: string
    x-enum-varnames:
    - Facebook
    - X
    - LinkedIn
    - Instagram
  publisher.PublishPostInfo:
    properties:
      media:
        items:
          $ref: '#/definitions/media.DownloadMetaData'
        type: array
      post:
        $ref: '#/definitions/post.PublishPost'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for OpenCM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OpenCM API
  version: "1.0"
paths:
...
  /publishers:
    get:
      consumes:
      - application/json
      description: Get available social networks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/publisher.Platform'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get available social networks
      tags:
      - publishers
  /publishers/{project_id}/{platform_id}/{user_id}/authenticate:
    post:
      consumes:
      - application/json
      description: Authenticate
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Platform ID
        in: path
        name: platform_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Authentication request
        in: body
        name: authentication
        required: true
        schema:
          $ref: '#/definitions/handlers.authenticationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Authenticate
      tags:
      - publishers
  /publishers/{project_id}/{post_id}/{platform_id}:
    post:
      consumes:
      - application/json
      description: Publish post to social network
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Social Network ID
        in: path
        name: social_network_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Publish post to social network
      tags:
      - publishers
  /publishers/{project_id}/{post_id}/{platform_id}/info:
    get:
      consumes:
      - application/json
      description: Get publish post info
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Social Network ID
        in: path
        name: social_network_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/publisher.PublishPostInfo'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get publish post info
      tags:
      - publishers
  /publishers/{project_id}/{post_id}/{platform_id}/validate:
    get:
      consumes:
      - application/json
      description: Validate post for social network
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Social Network ID
        in: path
        name: social_network_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Validate post for social network
      tags:
      - publishers
  /publishers/{project_id}/{post_id}/validate:
    get:
      consumes:
      - application/json
      description: Validate post for all assigned social networks
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Validate post for all assigned social networks
      tags:
      - publishers
...
schemes:
- http
- https
swagger: "2.0"